requirements.txt
asttokens==2.4.1
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
fonttools==4.47.0
ipython==8.20.0
jedi==0.19.1
kiwisolver==1.4.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mediapy==1.2.0
numpy==1.26.3
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
six==1.16.0
stack-data==0.6.3
traitlets==5.14.1
wcwidth==0.2.13
